skippinge existing files

+++ src/bin/vis_some.py
@@ -22,8 +22,10 @@ for layer, _ in tqdm(model.layer_info):
     save_dir = os.path.join(save_dir_root, model_name, layer)
     os.makedirs(save_dir, exist_ok=True)
     for i in range(10):
+        save_path = os.path.join(save_dir, str(i)+'.jpg')
+        if os.path.isfile(save_path):
+            continue
         vis_layer = layer + ':' + str(i)
         img = model.lucid(vis_layer, progress=False)
         img = cv2.cvtColor(img, cv2.COLOR_RGB2BGR)
-        save_path = os.path.join(save_dir, str(i)+'.jpg')
         cv2.imwrite(save_path, img)
